requirements.txt
albumentations==1.3.0
numpy==1.22.0
Pillow==9.0.1
pkg-resources==0.0.0
facenet-pytorch==2.5.3
iglovikov_helper_functions==0.0.53
people-segmentation==0.0.4
protobuf==3.15.0
six==1.12.0
tensorboardX==1.8
torch==1.2.0
torchvision==0.4.0
tqdm==4.35.0
opencv-python==4.4.0.46
imageio==2.9.0
